action component dynamically creates cards by type
